Create new components to easily compose a RewardCard
